requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=4.5
absl-py=0.11.0
alabaster=0.7.12
astunparse=1.6.3
attrs=20.3.0
babel=2.9.0
blas=1.1
bleach=1.5.0
ca-certificates=2021.10.8
cachetools=4.2.1
cdflib=0.3.20
certifi=2016.9.26
cffi=1.14.1
chardet=4.0.0
clearml=0.17.4
clearml-agent=0.17.1
cloudpickle=0.4.0
cudatoolkit=10.0.130
cycler=0.10.0
decorator=4.4.2
docutils=0.16
easydict=1.9
einops=0.3.2
exifread=2.1.2
ffnet=0.8.4
flatbuffers=1.12
fpdf2=2.1.0
freetype=2.10.2
funcsigs=1.0.2
furl=2.1.0
future=0.18.2
gast=0.3.3
google-auth=1.27.0
google-auth-oauthlib=0.4.2
google-pasta=0.2.0
gpxpy=1.1.2
grpcio=1.32.0
h5py=2.10.0
hdf5=1.10.4
html5lib=0.9999999
humanfriendly=9.1
idna=2.10
imageio=2.9.0
imagesize=1.2.0
importlib-metadata=1.7.0
intel-openmp=2019.4
intelpython=2021.3.0
jinja2=2.11.3
joblib=0.14.1
jpeg=9d
json-tricks=3.15.5
jsonschema=3.2.0
keras-preprocessing=1.1.2
kiwisolver=1.2.0
lcms2=2.11
ld_impl_linux-64=2.34
libblas=3.8.0
libcblas=3.8.0
libffi=3.2.1
libgcc-ng=9.3.0
libgfortran=3.0.0
libgfortran-ng=7.5.0
libgomp=9.3.0
liblapack=3.8.0
libopenblas=0.3.10
libpng=1.6.37
libprotobuf=3.13.0
libstdcxx-ng=9.3.0
libtiff=4.1.0
libwebp-base=1.1.0
loky=2.9.0
lz4-c=1.9.2
markdown=3.2.2
markupsafe=1.1.1
matplotlib=3.3.1
matplotlib-base=3.3.1
mkl=2018.0.3
mkl-service=2.1.0
ncurses=6.2
networkx=1.11
ninja=1.10.1
numpy=1.19.5
oauthlib=3.1.0
olefile=0.46
openblas=0.2.19
opencv-python=4.0.0.21
openssl=1.1.1l
opt-einsum=3.3.0
orderedmultidict=1.0.1
packaging=20.9
pandas=1.1.5
pathlib2=2.3.5
pillow=7.1.0
pip=20.2.2
prettytable=2.0.0
protobuf=3.13.0
psutil=5.8.0
py=1.10.0
pyasn1=0.4.8
pyasn1-modules=0.2.8
pycparser=2.20
pyfeatures=0.1.1
pygments=2.8.1
pyhocon=0.3.57
pyjwt=1.7.1
pyparsing=2.4.7
pyproj=1.9.5.1
pyrsistent=0.17.3
pytest=3.0.7
python=3.6.11
python-dateutil=2.6.0
python-graphviz=0.16
python_abi=3.6
pytorch=1.4.0
pytorch-model-summary=0.1.2
pytz=2021.1
pywavelets=1.1.1
pyyaml=5.1
readline=8.0
repoze-lru=0.7
requests=2.25.1
requests-file=1.5.1
requests-oauthlib=1.3.0
rsa=4.7.2
scikit-image=0.17.2
scikit-learn=0.24.2
scipy=1.4.1
seaborn=0.11.1
setuptools=49.6.0
six=1.15.0
sklearn=0.0
snowballstemmer=2.1.0
spacepy=0.2.2
sphinx=3.4.3
sphinxcontrib-applehelp=1.0.2
sphinxcontrib-devhelp=1.0.2
sphinxcontrib-htmlhelp=1.0.3
sphinxcontrib-jsmath=1.0.1
sphinxcontrib-qthelp=1.0.3
sphinxcontrib-serializinghtml=1.1.4
sqlite=3.33.0
tensorboard=2.4.1
tensorboard-plugin-wit=1.8.0
tensorboardx=1.4
tensorflow=2.4.1
tensorflow-estimator=2.4.0
tensorflow-tensorboard=0.4.0
termcolor=1.1.0
threadpoolctl=3.0.0
tifffile=2020.9.3
tk=8.6.10
torchvision=0.5.0
torchviz=0.0.1
tornado=6.0.4
tqdm=4.54.1
trains=0.16.4
typing=3.7.4.3
typing-extensions=3.7.4.3
urllib3=1.26.2
virtualenv=16.7.10
wcwidth=0.2.5
webencodings=0.5.1
werkzeug=1.0.1
wheel=0.35.1
wrapt=1.12.1
xmltodict=0.10.2
xz=5.2.5
zipp=3.1.0
zlib=1.2.11
zstd=1.4.5
